pyproject.toml
[project]
requires-python = "==3.12.*"

[tool.poetry]
name = "telegram-bot"
description = ""
version = "1.0.0"
authors = ["BlindfoldedSurgery <team@blindfolded.surgery>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.12"
httpx = "^0.27.0"
openai = "^1.16.1"
python-telegram-bot = "~21.5"


[tool.poetry.group.dev.dependencies]
mypy = "~1.11"
pytest = "^8.0.0"
ruff = "0.6.5"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# strict = true
# check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
check_untyped_defs = true
allow_untyped_defs = true
allow_incomplete_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "UP",
]
ignore = [
    # redundant-open-modes
    "UP015",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["telegram_bot"]
